change data type recovery duratio and resolution duration to decimal

+++ db/migrate/20150820053456_add_resolution_time_recoverty_time_to_incident_report.rb
@@ -1,7 +1,7 @@
 class AddResolutionTimeRecovertyTimeToIncidentReport < ActiveRecord::Migration
   def change
     add_column :incident_reports, :resolved_time, :datetime
-    add_column :incident_reports, :resolution_duration, :double
-    add_column :incident_reports, :recovery_duration, :double
+    add_column :incident_reports, :resolution_duration, :decimal
+    add_column :incident_reports, :recovery_duration, :decimal
   end
 end
